update: fix the dialoge no hidden again after kill the boss
