+++ Assets/Scripts/Dialgurmanage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff1cecd638a1b404ea53d860fcd0e512
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
