+++ Assets/Scripts/Jun/BossController.cs
@@ -31,8 +31,6 @@ public class BossController : MonoBehaviour
     [SerializeField]private TextMeshProUGUI bossHealthText;
     [SerializeField]private Image bossHealthBarImage;
     [SerializeField] private SpawnPrincess spawnPrincess;
-    [SerializeField] private GameObject dialogueBox;
-    [SerializeField] private TextMeshProUGUI dialogueText;
 
     private Vector2 arenaPos;
 
@@ -176,23 +174,11 @@ public class BossController : MonoBehaviour
                 bossHealthText.transform.parent.gameObject.SetActive(false);
                 Destroy(gameObject);
                 spawnPrincess.InstantiatePrincess();
-                ShowDialogue("Congratulations! You have defeated the boss and saved the princess.");
+                Dialgurmanage.i.ShowDialogue("Congratulations! You have defeated the boss and saved the princess.");
                 // Victory
             }
         }
     }
-    private void ShowDialogue(string message)
-    {
-        dialogueBox.SetActive(true);
-        dialogueText.fontSize = 15;
-        dialogueText.text = message;
-        StartCoroutine(HideDialogueAfterDelay(15f));
-    }
 
-    private IEnumerator HideDialogueAfterDelay(float delay)
-    {
-        yield return new WaitForSeconds(delay);
-        dialogueBox.SetActive(false);
-    }
 
 }
